Shifted MainScreen out of MainActivity.kt
