+++ app/src/main/java/com/theatretools/documa/MainActivity.kt
@@ -59,6 +59,7 @@ import com.theatretools.documa.data.ViewModelFactory
 import com.theatretools.documa.dataobjects.PresetItem
 import com.theatretools.documa.ui.theme.DocuMATheme
 import com.theatretools.documa.uiElements.DeviceSelector
+import com.theatretools.documa.uiElements.MainScreen
 import com.theatretools.documa.uiElements.PresetSelector
 import kotlinx.coroutines.launch
 
@@ -95,113 +96,3 @@ class MainActivity : ComponentActivity() {
     }
 }
 
-
-
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun MainScreen(
-    viewModel: AppViewModel, lifecycleOwner: LifecycleOwner,
-    addButtonAction: ()-> Unit,
-    editCardAction: (presetItem: PresetItem)-> Unit,
-) {
-
-
-    val scaffoldState = rememberBottomSheetScaffoldState()
-    val drawerState = rememberDrawerState(initialValue = DrawerValue.Closed)
-    val scrollState = rememberScrollState()
-    val scope = rememberCoroutineScope()
-
-    BottomSheetScaffold(
-        scaffoldState = scaffoldState,
-        sheetPeekHeight = 0.dp,
-        sheetContent = {
-            Column (
-            )
-            {
-                DeviceSelector(modifier = Modifier, viewModel, lifecycleOwner)
-                //Spacer(Modifier.height(8.dp))
-                Row (modifier = Modifier.horizontalScroll(scrollState)){
-                    OutlinedButton(onClick = addButtonAction , modifier = Modifier.padding(8.dp)) {
-                        Icon(Icons.Filled.Create, contentDescription = null)
-                        Text(text = AnnotatedString("Add"))
-                    }
-                    OutlinedButton(onClick = { /*TODO*/ }, modifier = Modifier.padding(8.dp)) {
-                        Icon(Icons.Filled.Search, contentDescription = null)
-                        Text(text = AnnotatedString("Show Plan"))
-                    }
-                    OutlinedButton(onClick = { /*TODO*/ }, modifier = Modifier.padding(8.dp)) {
-                        Icon(Icons.Default.List, contentDescription = null)
-                        Text(text = AnnotatedString("Show Groups"))
-                    }
-                }
-            }
-
-        }) { _ ->
-
-        val items = listOf(Icons.Default.Favorite, Icons.Default.Face, Icons.Default.Email)
-        val selectedItem = remember { mutableStateOf(items[0]) }
-        ModalNavigationDrawer(
-            drawerState = drawerState,
-            drawerContent = {
-                ModalDrawerSheet {
-                    //Spacer(Modifier.height(12.dp))
-                    items.forEach { item ->
-                        NavigationDrawerItem(
-                            icon = { Icon(item, contentDescription = null) },
-                            label = { Text(AnnotatedString(item.name)) },
-                            selected = item == selectedItem.value,
-                            onClick = {
-                                scope.launch { drawerState.close() }
-                                selectedItem.value = item
-                            },
-                            modifier = Modifier.padding(NavigationDrawerItemDefaults.ItemPadding)
-                        )
-                    } }
-            },
-        ) {
-            Scaffold(
-                modifier = Modifier.padding(8.dp),
-                topBar = {
-                    TopAppBar(
-                        title = { Text(AnnotatedString("")) },
-                        navigationIcon = {
-                            IconButton(
-                                onClick = {
-                                    scope.launch {
-                                        drawerState.apply {
-                                            if (isClosed) open() else close()
-                                        }
-                                    }
-                                }
-                            ) {
-                                Icon(Icons.Filled.Menu, contentDescription = null)
-                            }
-                        }
-                    )
-                },
-                floatingActionButtonPosition = FabPosition.End,
-                floatingActionButton = {
-                    ExtendedFloatingActionButton(
-                        onClick = {
-                            scope.launch { scaffoldState.bottomSheetState.expand()}
-                        }
-                    ) {
-                        Text(AnnotatedString("Select Fixture"))
-                        Icon(Icons.Filled.KeyboardArrowUp, contentDescription = "")
-                    }
-                },
-                content = { padding ->
-                    Box(/* TODO: refresh pull */) {
-                        PresetSelector(
-                            modifier = Modifier.padding(padding),
-                            viewModel = viewModel,
-                            lifecycleOwner = lifecycleOwner,
-                            action = editCardAction
-                        )
-                    }
-                    }
-            )
-        }
-    }
-}
